See comments below:
No npm install required!

// Command line imput should be as follows:
// node app.js <function> <filename>
// function can either be "basic", "cloze", or "read" (not case sensitive)
// filename example: "cards.txt" (if the file does not exist, it will be created)
// if no filename is provided, "cards.txt" will be used as a default filename

Uses node package "readline" instead of inquirer (I know, I'll use inquirer for the next one).  Once you have selected either "basic", "cloze", or "read" it will either prompt you to create basic flash cards, cloze-deletion flash cards, or read, parse, and display a formatted list of flash cards from a text file.

If you choose the basic cards, you will be prompted to give a question then an answer in the form of "front" and "back" of the flash cards.
If you choose the cloze cards, you will be prompted to give a full text and a cloze-deletion answer which will be removed from the full text to be displayed on the front of the card.  The Cloze